+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## Unreleased
 
+### Added
+
+- Added functionality to update the order of subcollections
+
 ## [10.0.0](https://github.com/dbmdz/cudami/releases/tag/10.0.0) - 2025-06-06
 
 ### Changed